requirements.txt
pandas
openai==0.28
google-cloud-aiplatform
pyyaml
